backend/requirements.txt
arrow==1.2.2
asgiref==3.5.2
binaryornot==0.4.4
boto3==1.24.26
botocore==1.27.26
certifi==2022.6.15
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cookiecutter==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.51.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
priority==2.0.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
wsproto==1.1.0
